requirements.txt
os
io
cv2
base64
pathlib 
ultralytics 
google.oauth2 
googleapiclient.discovery 
googleapiclient.http 
dotenv 